# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/": {"get": {"summary": "Index", "operationId": "index__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/browse": {"get": {"summary": "Browse Page", "operationId": "browse_page_browse_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/media": {"get": {"summary": "Api List Media", "operationId": "api_list_media_api_media_get", "parameters": [{"name": "media_type", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Media Type"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MediaListResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/browse": {"get": {"summary": "Api Browse", "operationId": "api_browse_api_browse_get", "parameters": [{"name": "media_type", "in": "query", "required": true, "schema": {"type": "string", "title": "Media Type"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BrowseResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/media-types": {"get": {"summary": "Api List Media Types", "operationId": "api_list_media_types_api_media_types_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MediaTypesResponse"}}}}}}}, "/api/item": {"get": {"summary": "Api Get Item", "operationId": "api_get_item_api_item_get", "parameters": [{"name": "media_type", "in": "query", "required": true, "schema": {"type": "string", "title": "Media Type"}}, {"name": "title", "in": "query", "required": true, "schema": {"type": "string", "title": "Title"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ItemDetailResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"summary": "Api Delete Item", "operationId": "api_delete_item_api_item_delete", "parameters": [{"name": "media_type", "in": "query", "required": true, "schema": {"type": "string", "title": "Media Type"}}, {"name": "title", "in": "query", "required": true, "schema": {"type": "string", "title": "Title"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MessageResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/item/rating": {"patch": {"summary": "Api Update Rating", "operationId": "api_update_rating_api_item_rating_patch", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateRatingRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MessageResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/chat": {"post": {"summary": "Chat", "operationId": "chat_api_chat_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"BrowseResponse": {"properties": {"media_type": {"type": "string", "title": "Media Type"}, "data": {"items": {"$ref": "#/components/schemas/EnrichedMediaItem"}, "type": "array", "title": "Data"}, "tmdb_supported": {"type": "boolean", "title": "Tmdb Supported"}, "tmdb_configured": {"type": "boolean", "title": "Tmdb Configured"}}, "type": "object", "required": ["media_type", "data", "tmdb_supported", "tmdb_configured"], "title": "BrowseResponse"}, "ChatRequest": {"properties": {"messages": {"items": {"additionalProperties": true, "type": "object"}, "type": "array", "title": "Messages"}}, "type": "object", "required": ["messages"], "title": "ChatRequest"}, "EnrichedMediaItem": {"properties": {"title": {"type": "string", "title": "Title"}, "genre": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Genre"}, "rating": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Rating"}, "notes": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Notes"}, "tmdb": {"anyOf": [{"$ref": "#/components/schemas/TmdbMeta"}, {"type": "null"}]}}, "type": "object", "required": ["title"], "title": "EnrichedMediaItem"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ItemDetail": {"properties": {"title": {"type": "string", "title": "Title"}, "genre": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Genre"}, "rating": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Rating"}, "notes": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Notes"}, "created_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Created At"}, "updated_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Updated At"}}, "type": "object", "required": ["title"], "title": "ItemDetail"}, "ItemDetailResponse": {"properties": {"media_type": {"type": "string", "title": "Media Type"}, "data": {"$ref": "#/components/schemas/ItemDetail"}}, "type": "object", "required": ["media_type", "data"], "title": "ItemDetailResponse"}, "MediaItem": {"properties": {"title": {"type": "string", "title": "Title"}, "genre": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Genre"}, "rating": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Rating"}, "notes": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Notes"}}, "type": "object", "required": ["title"], "title": "MediaItem"}, "MediaListResponse": {"properties": {"data": {"additionalProperties": {"items": {"$ref": "#/components/schemas/MediaItem"}, "type": "array"}, "type": "object", "title": "Data"}}, "type": "object", "required": ["data"], "title": "MediaListResponse"}, "MediaTypeInfo": {"properties": {"id": {"type": "integer", "title": "Id"}, "name": {"type": "string", "title": "Name"}}, "type": "object", "required": ["id", "name"], "title": "MediaTypeInfo"}, "MediaTypesResponse": {"properties": {"types": {"items": {"$ref": "#/components/schemas/MediaTypeInfo"}, "type": "array", "title": "Types"}}, "type": "object", "required": ["types"], "title": "MediaTypesResponse"}, "MessageResponse": {"properties": {"message": {"type": "string", "title": "Message"}}, "type": "object", "required": ["message"], "title": "MessageResponse"}, "TmdbMeta": {"properties": {"tmdb_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Tmdb Id"}, "title": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Title"}, "overview": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Overview"}, "poster_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Poster Url"}, "year": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Year"}, "tmdb_rating": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Tmdb Rating"}}, "type": "object", "title": "TmdbMeta"}, "UpdateRatingRequest": {"properties": {"media_type": {"type": "string", "title": "Media Type"}, "title": {"type": "string", "title": "Title"}, "rating": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Rating"}}, "type": "object", "required": ["media_type", "title"], "title": "UpdateRatingRequest"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}